+++ src/cli/util.js
@@ -15,7 +15,7 @@ const prettier = require("../../index");
 const errors = require("../common/errors");
 const resolver = require("../config/resolve-config");
 const constant = require("./constant");
-const coreOptions = require("../core/options");
+const coreOptions = require("../core/core-options");
 const optionsModule = require("../main/options");
 const optionsNormalizer = require("../main/options-normalizer");
 const thirdParty = require("../common/third-party");
