+++ src/core/core-options.js
